Addded Prereq class for compiling prerequisites
Moved the old compile_prereq_patterns and parse_prereqs functions into this class and changed names.
Now the patterns are compiled as a dict and stored as a class variable.
The parse_prereqs function is now the __init__ function of the class.
The class now has a parse_codes functions which takes in a parsable list of classes and outputs a list of lists of prereqs that are sufficient.